pyproject.toml
[tool.poetry]
name = "m_dynamodb_tools"
version = "0.1.0"
description = ""
authors = ["velo.yamigiku <velo.yamigiku@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
boto3 = "^1.35.87"
tqdm = "^4.67.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
